+++ Game/UI/UIPanelTrophyAchievements.cs
@@ -58,7 +58,7 @@ public class UIPanelTrophyAchievements : UIAppPanelBaseList {
 			
 			int i = 0;
 			
-			int totalPoints = 0;
+			double totalPoints = 0;
 			
 	        foreach(GameAchievement achievement in achievements) {
 				
@@ -84,7 +84,7 @@ public class UIPanelTrophyAchievements : UIAppPanelBaseList {
 				string points = "";
 				
 				if(completed) {
-					int currentPoints = achievement.data.points;
+                    double currentPoints = achievement.data.points;
 					totalPoints += currentPoints;
 					points = "+" + currentPoints.ToString();
